make due date a shared component
